requirements.txt
streamlit
openai-whisper
pydub
ffmpeg-python
torch
